test: migrate client double-Koin Compose UI tests to BisqComposeUiTestBase
