+++ apps/clientApp/src/androidUnitTest/kotlin/network/bisq/mobile/client/payment_accounts/presentation/create_payment_account/step2_payment_account_form/form/cash_deposit/CashDepositFormContentUiTest.kt
@@ -3,21 +3,12 @@ package network.bisq.mobile.client.payment_accounts.presentation.create_payment_
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.test.assertCountEquals
 import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.junit4.v2.createComposeRule
 import androidx.compose.ui.test.onAllNodesWithText
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTextInput
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import io.mockk.coEvery
 import io.mockk.mockk
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.test.StandardTestDispatcher
-import kotlinx.coroutines.test.advanceUntilIdle
-import kotlinx.coroutines.test.resetMain
-import kotlinx.coroutines.test.runTest
-import kotlinx.coroutines.test.setMain
 import network.bisq.mobile.client.common.test_utils.TestApplication
 import network.bisq.mobile.client.payment_accounts.domain.model.fiat.FiatPaymentMethod
 import network.bisq.mobile.client.payment_accounts.domain.model.fiat.cash_deposit.CreateCashDepositAccount
@@ -30,86 +21,41 @@ import network.bisq.mobile.client.payment_accounts.presentation.create_payment_a
 import network.bisq.mobile.data.replicated.account.payment_method.FiatPaymentRail
 import network.bisq.mobile.domain.model.account.create.CreatePaymentAccount
 import network.bisq.mobile.domain.model.account.fiat.FiatPaymentMethodChargebackRisk
-import network.bisq.mobile.domain.utils.CoroutineJobsManager
-import network.bisq.mobile.i18n.I18nSupport
 import network.bisq.mobile.i18n.i18n
-import network.bisq.mobile.presentation.common.ui.base.GlobalUiManager
 import network.bisq.mobile.presentation.common.ui.components.context.ExternalUrlOpener
 import network.bisq.mobile.presentation.common.ui.components.context.LocalExternalUrlOpener
-import network.bisq.mobile.presentation.common.ui.navigation.manager.NavigationManager
-import network.bisq.mobile.presentation.common.ui.theme.BisqTheme
-import network.bisq.mobile.presentation.common.ui.utils.LocalIsTest
 import network.bisq.mobile.presentation.main.MainPresenter
-import network.bisq.mobile.test.coroutines.TestCoroutineJobsManager
-import org.junit.After
-import org.junit.Before
-import org.junit.Rule
+import network.bisq.mobile.test.presentation.compose.BisqComposeUiTestBase
 import org.junit.Test
-import org.junit.runner.RunWith
-import org.koin.core.context.startKoin
-import org.koin.core.context.stopKoin
-import org.koin.dsl.module
 import org.robolectric.annotation.Config
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 
 @Config(application = TestApplication::class)
-@RunWith(AndroidJUnit4::class)
-@OptIn(ExperimentalCoroutinesApi::class)
-class CashDepositFormContentTest {
-    private val testDispatcher = StandardTestDispatcher()
-
-    @get:Rule
-    val composeTestRule = createComposeRule(effectContext = testDispatcher)
+class CashDepositFormContentUiTest : BisqComposeUiTestBase() {
     private lateinit var paymentAccountsServiceFacade: PaymentAccountsServiceFacade
     private lateinit var mainPresenter: MainPresenter
 
-    @Before
-    fun setup() {
-        Dispatchers.setMain(testDispatcher)
-        I18nSupport.setLanguage()
+    override fun setUpUiTest() {
+        super.setUpUiTest()
         paymentAccountsServiceFacade = mockk(relaxed = true)
         mainPresenter = mockk(relaxed = true)
-
-        runCatching { stopKoin() }
-        startKoin {
-            modules(
-                module {
-                    single<NavigationManager> { mockk(relaxed = true) }
-                    factory<CoroutineJobsManager> { TestCoroutineJobsManager(testDispatcher) }
-                    single<GlobalUiManager> { mockk(relaxed = true) }
-                },
-            )
-        }
     }
 
-    @After
-    fun tearDown() {
-        runCatching {
-            composeTestRule.setContent {}
-            composeTestRule.waitForIdle()
-        }
-        runCatching { stopKoin() }
-        Dispatchers.resetMain()
-    }
-
-    private fun setTestContent(
+    private fun setFormContent(
         presenter: CashDepositFormPresenter = CashDepositFormPresenter(paymentAccountsServiceFacade, mainPresenter),
         paymentMethod: FiatPaymentMethod = samplePaymentMethod(),
         onNavigateToNextScreen: (CreatePaymentAccount) -> Unit = {},
     ) {
-        composeTestRule.setContent {
+        setTestContent {
             CompositionLocalProvider(
-                LocalIsTest provides true,
                 LocalExternalUrlOpener provides ExternalUrlOpener { false },
             ) {
-                BisqTheme {
-                    CashDepositFormContent(
-                        presenter = presenter,
-                        onNavigateToNextScreen = onNavigateToNextScreen,
-                        paymentMethod = paymentMethod,
-                    )
-                }
+                CashDepositFormContent(
+                    presenter = presenter,
+                    onNavigateToNextScreen = onNavigateToNextScreen,
+                    paymentMethod = paymentMethod,
+                )
             }
         }
         dismissBackgroundDialogIfPresent()
@@ -129,7 +75,7 @@ class CashDepositFormContentTest {
 
     @Test
     fun `when rendered then country dropdown prompt is shown before country selection`() {
-        setTestContent()
+        setFormContent()
 
         composeTestRule.waitForIdle()
         composeTestRule.onNodeWithText("paymentAccounts.country".i18n()).assertIsDisplayed()
@@ -139,7 +85,7 @@ class CashDepositFormContentTest {
 
     @Test
     fun `when country dropdown search typed then countries are filtered`() {
-        setTestContent()
+        setFormContent()
 
         composeTestRule.onNodeWithText("paymentAccounts.createAccount.accountData.country.prompt".i18n()).performClick()
         composeTestRule.onNodeWithText("mobile.components.dropdown.searchPlaceholder".i18n()).performTextInput("ger")
@@ -153,7 +99,7 @@ class CashDepositFormContentTest {
     fun `when selected country details load then dynamic fields are shown`() {
         coEvery { paymentAccountsServiceFacade.getBankAccountCountryDetails("US") } returns Result.success(sampleCountryDetails())
 
-        setTestContent()
+        setFormContent()
         composeTestRule.onNodeWithText("paymentAccounts.createAccount.accountData.country.prompt".i18n()).performClick()
         composeTestRule.onNodeWithText("United States").performClick()
         composeTestRule.waitForIdle()
@@ -175,7 +121,7 @@ class CashDepositFormContentTest {
     fun `when country details fail then error state is shown`() {
         coEvery { paymentAccountsServiceFacade.getBankAccountCountryDetails("US") } returns Result.failure(RuntimeException("boom"))
 
-        setTestContent()
+        setFormContent()
         composeTestRule.onNodeWithText("paymentAccounts.createAccount.accountData.country.prompt".i18n()).performClick()
         composeTestRule.onNodeWithText("United States").performClick()
         composeTestRule.waitForIdle()
@@ -188,7 +134,7 @@ class CashDepositFormContentTest {
     fun `when currency dropdown search typed then currencies are filtered`() {
         coEvery { paymentAccountsServiceFacade.getBankAccountCountryDetails("US") } returns Result.success(sampleCountryDetails())
 
-        setTestContent()
+        setFormContent()
         selectUnitedStates()
         composeTestRule.onNodeWithText("paymentAccounts.createAccount.accountData.currency.prompt".i18n()).performClick()
         composeTestRule.onNodeWithText("mobile.components.dropdown.searchPlaceholder".i18n()).performTextInput("eur")
@@ -202,7 +148,7 @@ class CashDepositFormContentTest {
     fun `when text fields typed then visible inputs update through presenter`() {
         coEvery { paymentAccountsServiceFacade.getBankAccountCountryDetails("US") } returns Result.success(sampleCountryDetails())
 
-        setTestContent()
+        setFormContent()
         selectUnitedStates()
 
         composeTestRule
@@ -232,48 +178,45 @@ class CashDepositFormContentTest {
     }
 
     @Test
-    fun `when presenter emits navigate effect then navigation callback receives account`() =
-        runTest(testDispatcher) {
-            coEvery { paymentAccountsServiceFacade.getBankAccountCountryDetails("US") } returns Result.success(sampleCountryDetails())
-            val presenter = CashDepositFormPresenter(paymentAccountsServiceFacade, mainPresenter)
-            var navigatedAccount: CreatePaymentAccount? = null
+    fun `when presenter emits navigate effect then navigation callback receives account`() {
+        coEvery { paymentAccountsServiceFacade.getBankAccountCountryDetails("US") } returns Result.success(sampleCountryDetails())
+        val presenter = CashDepositFormPresenter(paymentAccountsServiceFacade, mainPresenter)
+        var navigatedAccount: CreatePaymentAccount? = null
 
-            setTestContent(
-                presenter = presenter,
-                onNavigateToNextScreen = { account -> navigatedAccount = account },
-            )
-            composeTestRule.waitForIdle()
-            presenter.onAction(CashDepositFormUiAction.OnCountrySelect(2))
-            advanceUntilIdle()
-            presenter.onAction(CashDepositFormUiAction.OnCurrencySelect(1))
-            presenter.onCommonAction(AccountFormUiAction.OnUniqueAccountNameChange("Cash Deposit Main"))
-            presenter.onAction(CashDepositFormUiAction.OnHolderNameChange(" Alice Doe "))
-            presenter.onAction(CashDepositFormUiAction.OnHolderIdChange(" ID-123 "))
-            presenter.onAction(CashDepositFormUiAction.OnBankNameChange(" Bisq Bank "))
-            presenter.onAction(CashDepositFormUiAction.OnBankIdChange(" BANKUS33 "))
-            presenter.onAction(CashDepositFormUiAction.OnBranchIdChange(" 001 "))
-            presenter.onAction(CashDepositFormUiAction.OnAccountNrChange(" 123456789 "))
-            presenter.onAction(CashDepositFormUiAction.OnBankAccountTypeSelect(BankAccountType.CHECKING))
-            presenter.onAction(CashDepositFormUiAction.OnNationalAccountIdChange(" NAT-123 "))
-            presenter.onAction(CashDepositFormUiAction.OnRequirementsChange(" Bring receipt "))
-            presenter.onCommonAction(AccountFormUiAction.OnNextClick)
-            advanceUntilIdle()
-            composeTestRule.waitForIdle()
+        setFormContent(
+            presenter = presenter,
+            onNavigateToNextScreen = { account -> navigatedAccount = account },
+        )
+        presenter.onAction(CashDepositFormUiAction.OnCountrySelect(2))
+        composeTestRule.waitForIdle()
+        presenter.onAction(CashDepositFormUiAction.OnCurrencySelect(1))
+        presenter.onCommonAction(AccountFormUiAction.OnUniqueAccountNameChange("Cash Deposit Main"))
+        presenter.onAction(CashDepositFormUiAction.OnHolderNameChange(" Alice Doe "))
+        presenter.onAction(CashDepositFormUiAction.OnHolderIdChange(" ID-123 "))
+        presenter.onAction(CashDepositFormUiAction.OnBankNameChange(" Bisq Bank "))
+        presenter.onAction(CashDepositFormUiAction.OnBankIdChange(" BANKUS33 "))
+        presenter.onAction(CashDepositFormUiAction.OnBranchIdChange(" 001 "))
+        presenter.onAction(CashDepositFormUiAction.OnAccountNrChange(" 123456789 "))
+        presenter.onAction(CashDepositFormUiAction.OnBankAccountTypeSelect(BankAccountType.CHECKING))
+        presenter.onAction(CashDepositFormUiAction.OnNationalAccountIdChange(" NAT-123 "))
+        presenter.onAction(CashDepositFormUiAction.OnRequirementsChange(" Bring receipt "))
+        presenter.onCommonAction(AccountFormUiAction.OnNextClick)
+        composeTestRule.waitForIdle()
 
-            val account = assertNotNull(navigatedAccount) as CreateCashDepositAccount
-            assertEquals("Cash Deposit Main", account.accountName)
-            assertEquals("US", account.accountPayload.selectedCountryCode)
-            assertEquals("GBP", account.accountPayload.selectedCurrencyCode)
-            assertEquals("Alice Doe", account.accountPayload.holderName)
-            assertEquals("ID-123", account.accountPayload.holderId)
-            assertEquals("Bisq Bank", account.accountPayload.bankName)
-            assertEquals("BANKUS33", account.accountPayload.bankId)
-            assertEquals("001", account.accountPayload.branchId)
-            assertEquals("123456789", account.accountPayload.accountNr)
-            assertEquals(BankAccountType.CHECKING, account.accountPayload.bankAccountType)
-            assertEquals("NAT-123", account.accountPayload.nationalAccountId)
-            assertEquals("Bring receipt", account.accountPayload.requirements)
-        }
+        val account = assertNotNull(navigatedAccount) as CreateCashDepositAccount
+        assertEquals("Cash Deposit Main", account.accountName)
+        assertEquals("US", account.accountPayload.selectedCountryCode)
+        assertEquals("GBP", account.accountPayload.selectedCurrencyCode)
+        assertEquals("Alice Doe", account.accountPayload.holderName)
+        assertEquals("ID-123", account.accountPayload.holderId)
+        assertEquals("Bisq Bank", account.accountPayload.bankName)
+        assertEquals("BANKUS33", account.accountPayload.bankId)
+        assertEquals("001", account.accountPayload.branchId)
+        assertEquals("123456789", account.accountPayload.accountNr)
+        assertEquals(BankAccountType.CHECKING, account.accountPayload.bankAccountType)
+        assertEquals("NAT-123", account.accountPayload.nationalAccountId)
+        assertEquals("Bring receipt", account.accountPayload.requirements)
+    }
 
     private fun selectUnitedStates() {
         composeTestRule.onNodeWithText("paymentAccounts.createAccount.accountData.country.prompt".i18n()).performClick()
