+++ apps/clientApp/src/androidUnitTest/kotlin/network/bisq/mobile/client/payment_accounts/presentation/create_payment_account/step2_payment_account_form/form/other_crypto/OtherCryptoFormContentUiTest.kt
@@ -1,110 +1,54 @@
 package network.bisq.mobile.client.payment_accounts.presentation.create_payment_account.step2_payment_account_form.form.other_crypto
 
-import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.test.assertCountEquals
 import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.junit4.v2.createComposeRule
 import androidx.compose.ui.test.onAllNodesWithText
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTextInput
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import io.mockk.mockk
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.test.StandardTestDispatcher
-import kotlinx.coroutines.test.advanceUntilIdle
-import kotlinx.coroutines.test.resetMain
-import kotlinx.coroutines.test.runTest
-import kotlinx.coroutines.test.setMain
 import network.bisq.mobile.client.common.test_utils.TestApplication
 import network.bisq.mobile.client.payment_accounts.domain.model.crypto.CryptoPaymentMethod
 import network.bisq.mobile.client.payment_accounts.domain.model.crypto.other_crypto.CreateOtherCryptoAssetAccount
 import network.bisq.mobile.client.payment_accounts.presentation.create_payment_account.step2_payment_account_form.form.AccountFormUiAction
 import network.bisq.mobile.client.payment_accounts.presentation.create_payment_account.step2_payment_account_form.form.crypto.CryptoAccountFormUiAction
 import network.bisq.mobile.domain.model.account.create.CreatePaymentAccount
-import network.bisq.mobile.domain.utils.CoroutineJobsManager
-import network.bisq.mobile.i18n.I18nSupport
 import network.bisq.mobile.i18n.i18n
-import network.bisq.mobile.presentation.common.ui.base.GlobalUiManager
-import network.bisq.mobile.presentation.common.ui.navigation.manager.NavigationManager
-import network.bisq.mobile.presentation.common.ui.theme.BisqTheme
 import network.bisq.mobile.presentation.common.ui.utils.EMPTY_STRING
-import network.bisq.mobile.presentation.common.ui.utils.LocalIsTest
 import network.bisq.mobile.presentation.main.MainPresenter
-import network.bisq.mobile.test.coroutines.TestCoroutineJobsManager
-import org.junit.After
-import org.junit.Before
-import org.junit.Rule
+import network.bisq.mobile.test.presentation.compose.BisqComposeUiTestBase
 import org.junit.Test
-import org.junit.runner.RunWith
-import org.koin.core.context.startKoin
-import org.koin.core.context.stopKoin
-import org.koin.dsl.module
 import org.robolectric.annotation.Config
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 
 @Config(application = TestApplication::class)
-@RunWith(AndroidJUnit4::class)
-@OptIn(ExperimentalCoroutinesApi::class)
-class OtherCryptoFormContentTest {
-    private val testDispatcher = StandardTestDispatcher()
-
-    @get:Rule
-    val composeTestRule = createComposeRule(effectContext = testDispatcher)
+class OtherCryptoFormContentUiTest : BisqComposeUiTestBase() {
     private lateinit var mainPresenter: MainPresenter
 
-    @Before
-    fun setup() {
-        Dispatchers.setMain(testDispatcher)
-        I18nSupport.setLanguage()
+    override fun setUpUiTest() {
+        super.setUpUiTest()
         mainPresenter = mockk(relaxed = true)
-
-        runCatching { stopKoin() }
-        startKoin {
-            modules(
-                module {
-                    single<NavigationManager> { mockk(relaxed = true) }
-                    factory<CoroutineJobsManager> { TestCoroutineJobsManager(testDispatcher) }
-                    single<GlobalUiManager> { mockk(relaxed = true) }
-                },
-            )
-        }
     }
 
-    @After
-    fun tearDown() {
-        try {
-            stopKoin()
-        } finally {
-            Dispatchers.resetMain()
-        }
-    }
-
-    private fun setTestContent(
+    private fun setFormContent(
         presenter: OtherCryptoFormPresenter = OtherCryptoFormPresenter(mainPresenter),
         paymentMethod: CryptoPaymentMethod = samplePaymentMethod(supportAutoConf = false),
         onNavigateToNextScreen: (CreatePaymentAccount) -> Unit = {},
     ) {
-        composeTestRule.setContent {
-            CompositionLocalProvider(LocalIsTest provides true) {
-                BisqTheme {
-                    OtherCryptoFormContent(
-                        presenter = presenter,
-                        paymentMethod = paymentMethod,
-                        onNavigateToNextScreen = onNavigateToNextScreen,
-                    )
-                }
-            }
+        setTestContent {
+            OtherCryptoFormContent(
+                presenter = presenter,
+                paymentMethod = paymentMethod,
+                onNavigateToNextScreen = onNavigateToNextScreen,
+            )
         }
     }
 
     @Test
     fun `when rendered then address and instant controls are shown`() {
-        setTestContent(paymentMethod = samplePaymentMethod(supportAutoConf = false))
+        setFormContent(paymentMethod = samplePaymentMethod(supportAutoConf = false))
 
-        composeTestRule.waitForIdle()
         composeTestRule
             .onNodeWithText("paymentAccounts.crypto.address.address".i18n())
             .assertIsDisplayed()
@@ -115,9 +59,8 @@ class OtherCryptoFormContentTest {
 
     @Test
     fun `when payment method supports auto conf then auto conf switch is shown`() {
-        setTestContent(paymentMethod = samplePaymentMethod(supportAutoConf = true))
+        setFormContent(paymentMethod = samplePaymentMethod(supportAutoConf = true))
 
-        composeTestRule.waitForIdle()
         composeTestRule
             .onNodeWithText("paymentAccounts.crypto.address.autoConf.use".i18n())
             .assertIsDisplayed()
@@ -125,9 +68,8 @@ class OtherCryptoFormContentTest {
 
     @Test
     fun `when payment method does not support auto conf then auto conf controls are hidden`() {
-        setTestContent(paymentMethod = samplePaymentMethod(supportAutoConf = false))
+        setFormContent(paymentMethod = samplePaymentMethod(supportAutoConf = false))
 
-        composeTestRule.waitForIdle()
         composeTestRule
             .onAllNodesWithText("paymentAccounts.crypto.address.autoConf.use".i18n())
             .assertCountEquals(0)
@@ -139,9 +81,8 @@ class OtherCryptoFormContentTest {
     @Test
     fun `when address typed then visible input updates through presenter`() {
         val typedAddress = "0xFEEDBEEF"
-        setTestContent(paymentMethod = samplePaymentMethod(supportAutoConf = false))
+        setFormContent(paymentMethod = samplePaymentMethod(supportAutoConf = false))
 
-        composeTestRule.waitForIdle()
         composeTestRule
             .onNodeWithText("paymentAccounts.crypto.address.address.prompt".i18n())
             .performTextInput(typedAddress)
@@ -152,9 +93,8 @@ class OtherCryptoFormContentTest {
 
     @Test
     fun `when auto conf supported and clicked then auto conf fields are shown`() {
-        setTestContent(paymentMethod = samplePaymentMethod(supportAutoConf = true))
+        setFormContent(paymentMethod = samplePaymentMethod(supportAutoConf = true))
 
-        composeTestRule.waitForIdle()
         composeTestRule
             .onNodeWithText("paymentAccounts.crypto.address.autoConf.use".i18n())
             .performClick()
@@ -174,9 +114,8 @@ class OtherCryptoFormContentTest {
     @Test
     fun `when auto conf field typed then visible input updates through presenter`() {
         val confirmations = "3"
-        setTestContent(paymentMethod = samplePaymentMethod(supportAutoConf = true))
+        setFormContent(paymentMethod = samplePaymentMethod(supportAutoConf = true))
 
-        composeTestRule.waitForIdle()
         composeTestRule
             .onNodeWithText("paymentAccounts.crypto.address.autoConf.use".i18n())
             .performClick()
@@ -190,30 +129,27 @@ class OtherCryptoFormContentTest {
     }
 
     @Test
-    fun `when presenter emits navigate effect then navigation callback is invoked`() =
-        runTest(testDispatcher) {
-            val presenter = OtherCryptoFormPresenter(mainPresenter)
-            val paymentMethod = samplePaymentMethod(supportAutoConf = false)
-            var navigatedAccount: CreatePaymentAccount? = null
+    fun `when presenter emits navigate effect then navigation callback is invoked`() {
+        val presenter = OtherCryptoFormPresenter(mainPresenter)
+        val paymentMethod = samplePaymentMethod(supportAutoConf = false)
+        var navigatedAccount: CreatePaymentAccount? = null
 
-            setTestContent(
-                presenter = presenter,
-                paymentMethod = paymentMethod,
-                onNavigateToNextScreen = { account -> navigatedAccount = account },
-            )
-            composeTestRule.waitForIdle()
+        setFormContent(
+            presenter = presenter,
+            paymentMethod = paymentMethod,
+            onNavigateToNextScreen = { account -> navigatedAccount = account },
+        )
 
-            presenter.onCommonAction(AccountFormUiAction.OnUniqueAccountNameChange("ETH Account"))
-            presenter.onCryptoCommonAction(CryptoAccountFormUiAction.OnAddressChange("0xABC123"))
-            presenter.onCommonAction(AccountFormUiAction.OnNextClick)
-            advanceUntilIdle()
+        presenter.onCommonAction(AccountFormUiAction.OnUniqueAccountNameChange("ETH Account"))
+        presenter.onCryptoCommonAction(CryptoAccountFormUiAction.OnAddressChange("0xABC123"))
+        presenter.onCommonAction(AccountFormUiAction.OnNextClick)
+        composeTestRule.waitForIdle()
 
-            composeTestRule.waitForIdle()
-            val account = assertNotNull(navigatedAccount) as CreateOtherCryptoAssetAccount
-            assertEquals("ETH Account", account.accountName)
-            assertEquals("0xABC123", account.accountPayload.address)
-            assertEquals("ETH", account.accountPayload.currencyCode)
-        }
+        val account = assertNotNull(navigatedAccount) as CreateOtherCryptoAssetAccount
+        assertEquals("ETH Account", account.accountName)
+        assertEquals("0xABC123", account.accountPayload.address)
+        assertEquals("ETH", account.accountPayload.currencyCode)
+    }
 
     private fun samplePaymentMethod(supportAutoConf: Boolean): CryptoPaymentMethod =
         CryptoPaymentMethod(
