+++ apps/clientApp/src/androidUnitTest/kotlin/network/bisq/mobile/client/payment_accounts/presentation/create_payment_account/step2_payment_account_form/form/wise/WiseFormContentUiTest.kt
@@ -1,22 +1,12 @@
 package network.bisq.mobile.client.payment_accounts.presentation.create_payment_account.step2_payment_account_form.form.wise
 
-import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.test.assertCountEquals
 import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.junit4.v2.createComposeRule
 import androidx.compose.ui.test.onAllNodesWithText
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTextInput
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import io.mockk.mockk
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.test.StandardTestDispatcher
-import kotlinx.coroutines.test.advanceUntilIdle
-import kotlinx.coroutines.test.resetMain
-import kotlinx.coroutines.test.runTest
-import kotlinx.coroutines.test.setMain
 import network.bisq.mobile.client.common.test_utils.TestApplication
 import network.bisq.mobile.client.payment_accounts.domain.model.fiat.FiatPaymentMethod
 import network.bisq.mobile.client.payment_accounts.domain.model.fiat.common.country.Country
@@ -26,87 +16,41 @@ import network.bisq.mobile.client.payment_accounts.presentation.create_payment_a
 import network.bisq.mobile.data.replicated.account.payment_method.FiatPaymentRail
 import network.bisq.mobile.domain.model.account.create.CreatePaymentAccount
 import network.bisq.mobile.domain.model.account.fiat.FiatPaymentMethodChargebackRisk
-import network.bisq.mobile.domain.utils.CoroutineJobsManager
-import network.bisq.mobile.i18n.I18nSupport
 import network.bisq.mobile.i18n.i18n
-import network.bisq.mobile.presentation.common.ui.base.GlobalUiManager
-import network.bisq.mobile.presentation.common.ui.navigation.manager.NavigationManager
-import network.bisq.mobile.presentation.common.ui.theme.BisqTheme
-import network.bisq.mobile.presentation.common.ui.utils.LocalIsTest
 import network.bisq.mobile.presentation.main.MainPresenter
-import network.bisq.mobile.test.coroutines.TestCoroutineJobsManager
-import org.junit.After
-import org.junit.Before
-import org.junit.Rule
+import network.bisq.mobile.test.presentation.compose.BisqComposeUiTestBase
 import org.junit.Test
-import org.junit.runner.RunWith
-import org.koin.core.context.startKoin
-import org.koin.core.context.stopKoin
-import org.koin.dsl.module
 import org.robolectric.annotation.Config
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 
 @Config(application = TestApplication::class)
-@RunWith(AndroidJUnit4::class)
-@OptIn(ExperimentalCoroutinesApi::class)
-class WiseFormContentTest {
-    private val testDispatcher = StandardTestDispatcher()
-
-    @get:Rule
-    val composeTestRule = createComposeRule(effectContext = testDispatcher)
+class WiseFormContentUiTest : BisqComposeUiTestBase() {
     private lateinit var mainPresenter: MainPresenter
 
-    @Before
-    fun setup() {
-        Dispatchers.setMain(testDispatcher)
-        I18nSupport.setLanguage()
+    override fun setUpUiTest() {
+        super.setUpUiTest()
         mainPresenter = mockk(relaxed = true)
-
-        runCatching { stopKoin() }
-        startKoin {
-            modules(
-                module {
-                    single<NavigationManager> { mockk(relaxed = true) }
-                    factory<CoroutineJobsManager> { TestCoroutineJobsManager(testDispatcher) }
-                    single<GlobalUiManager> { mockk(relaxed = true) }
-                },
-            )
-        }
     }
 
-    @After
-    fun tearDown() {
-        try {
-            stopKoin()
-        } finally {
-            Dispatchers.resetMain()
-        }
-    }
-
-    private fun setTestContent(
+    private fun setFormContent(
         presenter: WiseFormPresenter = WiseFormPresenter(mainPresenter),
         paymentMethod: FiatPaymentMethod = samplePaymentMethod(),
         onNavigateToNextScreen: (CreatePaymentAccount) -> Unit = {},
     ) {
-        composeTestRule.setContent {
-            CompositionLocalProvider(LocalIsTest provides true) {
-                BisqTheme {
-                    WiseFormContent(
-                        presenter = presenter,
-                        onNavigateToNextScreen = onNavigateToNextScreen,
-                        paymentMethod = paymentMethod,
-                    )
-                }
-            }
+        setTestContent {
+            WiseFormContent(
+                presenter = presenter,
+                onNavigateToNextScreen = onNavigateToNextScreen,
+                paymentMethod = paymentMethod,
+            )
         }
     }
 
     @Test
     fun `when rendered then wise form fields and currency summary are shown`() {
-        setTestContent()
+        setFormContent()
 
-        composeTestRule.waitForIdle()
         composeTestRule.onNodeWithText("paymentAccounts.holderName".i18n()).assertIsDisplayed()
         composeTestRule.onNodeWithText("paymentAccounts.email".i18n()).assertIsDisplayed()
         composeTestRule.onNodeWithText("mobile.paymentAccounts.currencyPicker.allSelected".i18n(3)).assertIsDisplayed()
@@ -115,9 +59,8 @@ class WiseFormContentTest {
     @Test
     fun `when holder name field typed then visible input updates through presenter`() {
         val holderName = "Alice Doe"
-        setTestContent()
+        setFormContent()
 
-        composeTestRule.waitForIdle()
         composeTestRule
             .onNodeWithText(
                 "paymentAccounts.createAccount.prompt".i18n(
@@ -132,9 +75,8 @@ class WiseFormContentTest {
     @Test
     fun `when email field typed then visible input updates through presenter`() {
         val email = "alice@example.com"
-        setTestContent()
+        setFormContent()
 
-        composeTestRule.waitForIdle()
         composeTestRule
             .onNodeWithText(
                 "paymentAccounts.createAccount.prompt".i18n(
@@ -148,9 +90,8 @@ class WiseFormContentTest {
 
     @Test
     fun `when currency summary clicked then picker controls and currencies are shown`() {
-        setTestContent()
+        setFormContent()
 
-        composeTestRule.waitForIdle()
         composeTestRule
             .onNodeWithText("mobile.paymentAccounts.currencyPicker.allSelected".i18n(3))
             .performClick()
@@ -167,9 +108,8 @@ class WiseFormContentTest {
 
     @Test
     fun `when picker clear all clicked then currency summary updates through presenter`() {
-        setTestContent()
+        setFormContent()
 
-        composeTestRule.waitForIdle()
         composeTestRule
             .onNodeWithText("mobile.paymentAccounts.currencyPicker.allSelected".i18n(3))
             .performClick()
@@ -183,9 +123,8 @@ class WiseFormContentTest {
 
     @Test
     fun `when picker currency clicked then currency summary updates through presenter`() {
-        setTestContent()
+        setFormContent()
 
-        composeTestRule.waitForIdle()
         composeTestRule
             .onNodeWithText("mobile.paymentAccounts.currencyPicker.allSelected".i18n(3))
             .performClick()
@@ -199,9 +138,8 @@ class WiseFormContentTest {
 
     @Test
     fun `when picker search typed then list filters through presenter`() {
-        setTestContent()
+        setFormContent()
 
-        composeTestRule.waitForIdle()
         composeTestRule
             .onNodeWithText("mobile.paymentAccounts.currencyPicker.allSelected".i18n(3))
             .performClick()
@@ -215,9 +153,8 @@ class WiseFormContentTest {
 
     @Test
     fun `when picker search has no results then empty state is shown`() {
-        setTestContent()
+        setFormContent()
 
-        composeTestRule.waitForIdle()
         composeTestRule
             .onNodeWithText("mobile.paymentAccounts.currencyPicker.allSelected".i18n(3))
             .performClick()
@@ -229,31 +166,28 @@ class WiseFormContentTest {
     }
 
     @Test
-    fun `when presenter emits navigate effect then navigation callback receives account`() =
-        runTest(testDispatcher) {
-            val presenter = WiseFormPresenter(mainPresenter)
-            var navigatedAccount: CreatePaymentAccount? = null
+    fun `when presenter emits navigate effect then navigation callback receives account`() {
+        val presenter = WiseFormPresenter(mainPresenter)
+        var navigatedAccount: CreatePaymentAccount? = null
 
-            setTestContent(
-                presenter = presenter,
-                onNavigateToNextScreen = { account -> navigatedAccount = account },
-            )
-            composeTestRule.waitForIdle()
+        setFormContent(
+            presenter = presenter,
+            onNavigateToNextScreen = { account -> navigatedAccount = account },
+        )
 
-            presenter.onCommonAction(AccountFormUiAction.OnUniqueAccountNameChange("Wise Personal"))
-            presenter.onAction(WiseFormUiAction.OnHolderNameChange("John Doe"))
-            presenter.onAction(WiseFormUiAction.OnEmailChange("john@example.com"))
-            presenter.onAction(WiseFormUiAction.OnCurrencyToggle("GBP"))
-            presenter.onCommonAction(AccountFormUiAction.OnNextClick)
-            advanceUntilIdle()
+        presenter.onCommonAction(AccountFormUiAction.OnUniqueAccountNameChange("Wise Personal"))
+        presenter.onAction(WiseFormUiAction.OnHolderNameChange("John Doe"))
+        presenter.onAction(WiseFormUiAction.OnEmailChange("john@example.com"))
+        presenter.onAction(WiseFormUiAction.OnCurrencyToggle("GBP"))
+        presenter.onCommonAction(AccountFormUiAction.OnNextClick)
+        composeTestRule.waitForIdle()
 
-            composeTestRule.waitForIdle()
-            val account = assertNotNull(navigatedAccount) as CreateWiseAccount
-            assertEquals("Wise Personal", account.accountName)
-            assertEquals("John Doe", account.accountPayload.holderName)
-            assertEquals("john@example.com", account.accountPayload.email)
-            assertEquals(listOf("EUR", "USD"), account.accountPayload.selectedCurrencies.map { currency -> currency.code })
-        }
+        val account = assertNotNull(navigatedAccount) as CreateWiseAccount
+        assertEquals("Wise Personal", account.accountName)
+        assertEquals("John Doe", account.accountPayload.holderName)
+        assertEquals("john@example.com", account.accountPayload.email)
+        assertEquals(listOf("EUR", "USD"), account.accountPayload.selectedCurrencies.map { currency -> currency.code })
+    }
 
     private fun samplePaymentMethod(): FiatPaymentMethod =
         FiatPaymentMethod(
