+++ apps/clientApp/src/androidUnitTest/kotlin/network/bisq/mobile/client/payment_accounts/presentation/create_payment_account/step2_payment_account_form/form/zelle/ZelleFormContentUiTest.kt
@@ -3,108 +3,52 @@ package network.bisq.mobile.client.payment_accounts.presentation.create_payment_
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.test.assertCountEquals
 import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.junit4.v2.createComposeRule
 import androidx.compose.ui.test.onAllNodesWithText
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTextInput
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import io.mockk.mockk
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.test.StandardTestDispatcher
-import kotlinx.coroutines.test.advanceUntilIdle
-import kotlinx.coroutines.test.resetMain
-import kotlinx.coroutines.test.runTest
-import kotlinx.coroutines.test.setMain
 import network.bisq.mobile.client.common.test_utils.TestApplication
 import network.bisq.mobile.client.payment_accounts.domain.model.fiat.zelle.CreateZelleAccount
 import network.bisq.mobile.client.payment_accounts.presentation.create_payment_account.step2_payment_account_form.form.AccountFormUiAction
 import network.bisq.mobile.domain.model.account.create.CreatePaymentAccount
-import network.bisq.mobile.domain.utils.CoroutineJobsManager
-import network.bisq.mobile.i18n.I18nSupport
 import network.bisq.mobile.i18n.i18n
-import network.bisq.mobile.presentation.common.ui.base.GlobalUiManager
 import network.bisq.mobile.presentation.common.ui.components.context.ExternalUrlOpener
 import network.bisq.mobile.presentation.common.ui.components.context.LocalExternalUrlOpener
-import network.bisq.mobile.presentation.common.ui.navigation.manager.NavigationManager
-import network.bisq.mobile.presentation.common.ui.theme.BisqTheme
-import network.bisq.mobile.presentation.common.ui.utils.LocalIsTest
 import network.bisq.mobile.presentation.main.MainPresenter
-import network.bisq.mobile.test.coroutines.TestCoroutineJobsManager
-import org.junit.After
-import org.junit.Before
-import org.junit.Rule
+import network.bisq.mobile.test.presentation.compose.BisqComposeUiTestBase
 import org.junit.Test
-import org.junit.runner.RunWith
-import org.koin.core.context.startKoin
-import org.koin.core.context.stopKoin
-import org.koin.dsl.module
 import org.robolectric.annotation.Config
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 
 @Config(application = TestApplication::class)
-@RunWith(AndroidJUnit4::class)
-@OptIn(ExperimentalCoroutinesApi::class)
-class ZelleFormContentTest {
-    private val testDispatcher = StandardTestDispatcher()
-
-    @get:Rule
-    val composeTestRule = createComposeRule(effectContext = testDispatcher)
+class ZelleFormContentUiTest : BisqComposeUiTestBase() {
     private lateinit var mainPresenter: MainPresenter
 
-    @Before
-    fun setup() {
-        Dispatchers.setMain(testDispatcher)
-        I18nSupport.setLanguage()
+    override fun setUpUiTest() {
+        super.setUpUiTest()
         mainPresenter = mockk(relaxed = true)
-
-        runCatching { stopKoin() }
-        startKoin {
-            modules(
-                module {
-                    single<NavigationManager> { mockk(relaxed = true) }
-                    factory<CoroutineJobsManager> { TestCoroutineJobsManager(testDispatcher) }
-                    single<GlobalUiManager> { mockk(relaxed = true) }
-                },
-            )
-        }
     }
 
-    @After
-    fun tearDown() {
-        try {
-            stopKoin()
-        } finally {
-            Dispatchers.resetMain()
-        }
-    }
-
-    private fun setTestContent(
+    private fun setFormContent(
         presenter: ZelleFormPresenter = ZelleFormPresenter(mainPresenter),
         onNavigateToNextScreen: (CreatePaymentAccount) -> Unit = {},
     ) {
-        composeTestRule.setContent {
-            CompositionLocalProvider(
-                LocalIsTest provides true,
-                LocalExternalUrlOpener provides ExternalUrlOpener { true },
-            ) {
-                BisqTheme {
-                    ZelleFormContent(
-                        presenter = presenter,
-                        onNavigateToNextScreen = onNavigateToNextScreen,
-                    )
-                }
+        setTestContent {
+            CompositionLocalProvider(LocalExternalUrlOpener provides ExternalUrlOpener { true }) {
+                ZelleFormContent(
+                    presenter = presenter,
+                    onNavigateToNextScreen = onNavigateToNextScreen,
+                )
             }
         }
     }
 
     @Test
     fun `when rendered then zelle form fields and background dialog are shown`() {
-        setTestContent()
+        setFormContent()
 
-        composeTestRule.waitForIdle()
         composeTestRule.onNodeWithText("paymentAccounts.holderName".i18n()).assertIsDisplayed()
         composeTestRule.onNodeWithText("paymentAccounts.emailOrMobileNr".i18n()).assertIsDisplayed()
         composeTestRule
@@ -115,9 +59,8 @@ class ZelleFormContentTest {
 
     @Test
     fun `when dismissing background dialog then it is hidden`() {
-        setTestContent()
+        setFormContent()
 
-        composeTestRule.waitForIdle()
         composeTestRule.onNodeWithText("action.iUnderstand".i18n()).performClick()
 
         composeTestRule.waitForIdle()
@@ -130,7 +73,7 @@ class ZelleFormContentTest {
     @Test
     fun `when holder name field typed then visible input updates through presenter`() {
         val holderName = "Alice Doe"
-        setTestContent()
+        setFormContent()
         composeTestRule.onNodeWithText("action.iUnderstand".i18n()).performClick()
 
         composeTestRule.waitForIdle()
@@ -148,7 +91,7 @@ class ZelleFormContentTest {
     @Test
     fun `when email mobile field typed then visible input updates through presenter`() {
         val emailOrMobile = "alice@example.com"
-        setTestContent()
+        setFormContent()
         composeTestRule.onNodeWithText("action.iUnderstand".i18n()).performClick()
 
         composeTestRule.waitForIdle()
@@ -164,27 +107,24 @@ class ZelleFormContentTest {
     }
 
     @Test
-    fun `when presenter emits navigate effect then navigation callback receives account`() =
-        runTest(testDispatcher) {
-            val presenter = ZelleFormPresenter(mainPresenter)
-            var navigatedAccount: CreatePaymentAccount? = null
+    fun `when presenter emits navigate effect then navigation callback receives account`() {
+        val presenter = ZelleFormPresenter(mainPresenter)
+        var navigatedAccount: CreatePaymentAccount? = null
 
-            setTestContent(
-                presenter = presenter,
-                onNavigateToNextScreen = { account -> navigatedAccount = account },
-            )
-            composeTestRule.waitForIdle()
+        setFormContent(
+            presenter = presenter,
+            onNavigateToNextScreen = { account -> navigatedAccount = account },
+        )
 
-            presenter.onCommonAction(AccountFormUiAction.OnUniqueAccountNameChange("Zelle Personal"))
-            presenter.onAction(ZelleFormUiAction.OnHolderNameChange("John Doe"))
-            presenter.onAction(ZelleFormUiAction.OnEmailOrMobileNrChange("john@example.com"))
-            presenter.onCommonAction(AccountFormUiAction.OnNextClick)
-            advanceUntilIdle()
+        presenter.onCommonAction(AccountFormUiAction.OnUniqueAccountNameChange("Zelle Personal"))
+        presenter.onAction(ZelleFormUiAction.OnHolderNameChange("John Doe"))
+        presenter.onAction(ZelleFormUiAction.OnEmailOrMobileNrChange("john@example.com"))
+        presenter.onCommonAction(AccountFormUiAction.OnNextClick)
+        composeTestRule.waitForIdle()
 
-            composeTestRule.waitForIdle()
-            val account = assertNotNull(navigatedAccount) as CreateZelleAccount
-            assertEquals("Zelle Personal", account.accountName)
-            assertEquals("John Doe", account.accountPayload.holderName)
-            assertEquals("john@example.com", account.accountPayload.emailOrMobileNr)
-        }
+        val account = assertNotNull(navigatedAccount) as CreateZelleAccount
+        assertEquals("Zelle Personal", account.accountName)
+        assertEquals("John Doe", account.accountPayload.holderName)
+        assertEquals("john@example.com", account.accountPayload.emailOrMobileNr)
+    }
 }
